Fixing a crash an and UI issue:
#44 - NullPointerException when accessing screenUserControls using an index if we expect the list to exist. I made sure the list exists in advance. This change meant that the rules to fill the list were slightly changed, which I addressed and added comments to explain why.

PrefsByScreenUserControl does not have the correct size to show all the buttons at the same time when inside PreferencesForm. I am facing this issue in a high resolution screen in Windows 10. I fixed it by making sure the sizes are exact. I also made sure that the sub-form would stay in the same place when switching tabs.